project setup trivialities
This repository is an experiment in something loosely inspired by
literate programming: the project's commit history is constantly
rewritten to hopefully form an accessible introduction to the code
base.  The commit messages are transformed into a coherent(?)
document with `s/render.sh`.

We'll just have to see if that approach is fruitful and sustainable.