ground/variable: lightweight value type with opaque identity
Variables in first-order logic predicates have no specific value, only
an identity, so there's a lot of design freedom.

We don't expect to observe too many variables (we're mostly interested
in human-readable proofs), so we will simply represent variables as
32-bit integers, and generate fresh variables with a gensym counter.

That's about as simple and compact as it gets, for a first cut.

We'll also derive a `Ord` on these variables because it's convenient
to be able to use ordered containers like BTreeSet or, later, sorted
vectors.