ground/record: store individual facts/captures as &[Variable]
We expect to store facts and derived information in homogeneous
containers, where each element represents one instance of a specific
form of information (e.g., "this predicate is known to be true for al
the tuples in the collection").  If variables are the value we'll most
commonly manipulate, records of variables come immediately after.

For now, we use boxed slices instead of vectors in order to shave a
`usize` of heap footprint.  Eventually, we'll probably want to store
a fingerprint inline, to speed up hashing and comparison, or maybe
implement some small object optimisation (we expect a lot of small
records, rather than a few large ones).  That can all wait.