unification/metavariable: matching variables
In order to know what a collection of captures means, and to, e.g.,
make two such collections compatibles, we need a way to refer to
$\forall$-variables in antecedents.  Our ground values are
(tuples of) `Variable`s, so we'll call those `MetaVar`iables.

The only thing we'll use in these `MetaVar`s is their identity, but
it's also useful to have a printed representation for debugging.
That's why we'll generate unique identifiers by incrementing a `u64`,
but still adorn them with a `name` that's only used for debug
formatting.  The overhead of that string should be negligible, since
we don't store `MetaVar`s directly in collections, but only alongside
them.

We also don't want to use the `name` for identity, because that would
open us up to all sorts of nasty scoping bugs.  Give our first-order
sequents, it really makes sense to handle all scope issues in the
parser.

`MetaVar`s are also important for join ordering: they represent
attributes for equijoin (or antijoin) keys.  We'll make it easy to do
something not too horrible with sequentially allocated ids: `MetaVar`s
that are defined earlier in clauses will compare lesser, and should
probably be joined later.