unification/metavariable: matching variables
This is how we'll pair what values (variable) goes with what pattern
variable (metevariable).

Metavariable are also important for join ordering: they represent
attributes for equijoining (or antijoins).  We'll make it easy to
do something not too horrible with linearly sequential allocation
ids: metavariables that are defined earlier in clauses will compare
lesser, and should probably be joined later.

There also shouldn't be too many of them, and metavariables shouldn't
be Clone()d in inner loops, so we annotate each one with a string
description (which is only used for pretty printing).