unification/pattern: new function instantiate_template
Most instantiations shouldn't be on the fast path, so we can use a
straightforward interface: accept a template (slice of
`pattern::Element`) and an environment map from MetaVar to variable,
and return a Fact record.

We expect this function to be called with partial environments, with
missing captures backfilled after the fact, to represent existentials.