.## matching: pattern matching in bulk
We know how to check if a `Fact` tuple matches a pattern, and how to
extract a `Capture` as a witness of that match, if so.  The `matching`
modules takes this tuple-at-a-time logic, and uses it to find all
matches in *collections* of `Fact`s.

If we only had to deal with left-hand side antecendents of the form
`p(x, y)`, this would be a trivial `filter_map` task.  However, we
wish to handle antecedents that span multiple predicates,
e.g. $\forall x, y. p(x, y) \wedge q(y, y)$, or even antecedents
that match against the same predicate multiple times, e.g.,
$\forall x, y, z. p(x, y) \wedge p(y, z)$.

Finding matches for these more complex antecedents is equivalent to
executing a (self-)join.  We'll want to separately plan the join, and
lower a join plan to a differential dataflow expression graph: DD
plays a lot of type-level tricks to forbid escaping, so we try to do
work in advance when we can.  The result is easier to test, read, and
debug.