execution/split_container: thin wrapper around DD containers
Our core is fairly abstracted away from the level type safety that can
be realistically offered by a Rust library.  We instead dynamically
look for shape mismatches at the level of containers (`Collection`s and
`Variable`s).

We know we'll only ever collections of facts (axioms or derived), or
temporary collections of captures.  Each case can be described
differently (facts only have an arity, captures are associated with a
parallel list of metavariables).  Use a different type for each.