matching/plan: add Project nodes
We'll sometimes capture useless metavars, and often want to get rid of
metavars that are only relevant for a join, and not afterwards.
Projections are parameterised on an input shape and a list of metavars
we wish to keep; they accept captures matching the input shape, and
yield captures matching the "kept" list.