matching/plan: add Join nodes
Joins are parameterised on a join key (a list of metavars), a list of
metavars that are still interesting after the join, and a list of
input shapes.

A join node extract variable values corresponding to the "key"
metavars in each input, matches up the corresponding values, and
yields the variable values for the "kept" metavars, for each match.

The natural join is a convenience layer on top of the join node that
matches inputs on all metavars that exist in all nodes, and yields
all metavars that are uniquely defined (either in the join key, or
only appears in one input).

For now, it's only possible to construct joins of exactly two inputs;
this also simplifies the definition of a natural join.