matching/plan, execution/lower_plan: add Antijoin nodes
A key peculiarity of coherent logic is that it can implement *some*
existentials in consequents, by noticing when a consequent is already
satisfied, and not considering corresponding instances of a sequent.

We will implement this negative matching by matching on a sequent's
consequents, and antijoining the result from the antecendents'
potential matches.