execution/enumerate_candidates: find instantiation candidates
Non-trivial sequents (with existentials and/or multiple conjunctions
on the right-hand side) can't be handled directly as part of a datalog
fixpoint: there may well be no (finite) fixpoint.

Instead, enumerate everything that could be instantiated, and let a
higher-level process decide what to feed back in the input collections
of facts.