deduce: hook up trivial and enumerate_candidates together
The function `deduction_graph_from_sequents` accepts a timely scope, a
vector of trivial and another of search sequents, and constructs a
dataflow graph to propagate facts to trivial sequents, and hooks that
up to another dataflow graph that finds potential instances for search
sequents.  That dataflow graph is constructed in the timely scope, but
hidden; the function returns only the graph's input and output nodes.