.# Satumerator = SAT + Enumerator
or how to prove you really tried everything.

Discrete search problems, like theorem proving, often resort to
implicit tree enumeration not because it's a good strategy, but
because it's a simple way to guarantee the search is exhaustive.

SAT solvers have shown us that traversing an implicit tree is a bad
strategy for hard, but tractable, problems.  We really want the
ability to generalise whenever we close a branch, and thus avoid
exploring similar neighbours; we would also prefer to be able to
switch up the tree traversal order dynamically, without losing all the
progress we've made until then.

Satumerator uses `CryptoMiniSat` to expose a high level "nogood"
learning interface for arbitrary, potentially infinite, discrete
search problems.  The approach is based on
[Jannis Harder](https://jix.one/)'s
[model minimisation demo](https://github.com/jix/minimal_models_example);
we feed regular "nogood" constraints to one solver, and a
[Tseitin-transformed](https://en.wikipedia.org/wiki/Tseytin_transformation)
version of the same constraints to another.

This lets us ask the first solver whether we're done searching, and
get a witness of a "gap" in the nogoods if we still have work to do.
Given this witness (or any partial assignment), we then use the second
solver to detect whether there exists an extension of the assignment
that is forbidden by a nogood; our goal is to generate small partial
assignments such that we can arbitrarily dive down from that initial
state and not find ourselves exploring a region of the search space
that's already covered by a nogood.

In addition to regular "nogood" clauses, Satumerator also supports
domain constraints of the form "exactly one of x, y, z must be true
in any valid assignment."  These constraints let us express discrete
choices, for example, when one of three patterns may be instantiated
in a sequent.

Given these constraints, we can also generate "gap" witness that only
fix variables to true; that's useful in search problems, like our
coherent logic prover, where we can generate a fact from `x = true`,
but not really do anything with `y = false` (do not instantiate a
given pattern, without knowing whether to instantiate any of the other
options, or which).