.## `PositiveSatumerator`: Satumerate only with truths
The search problem in our coherent logic prover is more naturally seen
as a exploring a rose-tree with multi-way branches than a binary
tree.  It doesn't really mean much to know that a given consequent
should *not* be instantiated; we really want to know which of the
others might be interesting.

The `PositiveSatumerator` exposes a more restricted interface, where
every decision variable (consequent to instantiate or not instantiate)
must be bound by an "exactly one of" cardinality constraint, and all
"nogoods" must only include variables set to true.  In fact, the
interface only accepts a set of variables as nogoods, without any
associated truth value (the variables must be true in the forbidden
region), and represents partial assignments as vectors of
variables (atoms), with an implicit "true" value.