.## construct a "complemented" checker SAT state
We currently have a SAT state to which we can add nogoods and domain
constraints, in order to ask if we're done searching yet.

We also want to improve partial assignment by removing redundant
assignments: the gap (inexhaustiveness) witness we received from the
currently existing SAT state gives a value to all variables declared
so far.  While true, this assignment is too eager to be useful: we can
hopefully find a smaller partial assignment such that no (valid)
extension of that smaller assignment is covered by the nogoods.

The usual trick to answer `forall x, p(x)` in SAT is to flip the
expression and work with `exists x, !p(x)`: if there is no `x` that
falsifies `p(x)`, then `p(x)` holds for all `x`.  That's what we'll do
here...

But that requires a SAT formulation for `!p(x)`, and, while it's easy
to convert nogood constraints to CNF, the complement can easily
explode.  We'll use Tseitin's encoding trick: rather than directly
adding a constraint for a given expression, we'll add an output
variable and constrain that variable to take the same truth value as
the expression. At the end of it all, we can force the output variable
to false (with a temporary assumption), and that's equivalent to
forcing the expression to be violated.

Given a partial assignment, we can ask SAT to extend it while
respecting the domain constraints, and while violating at least one of
the nogoods.  What we're hoping for is a negative (UNSAT) answer: this
means the partial assignment is such that any extension isn't
forbidden by the nogoods.

When we get UNSAT, we can ask CryptoMiniSat for a conflict set from
our initial assumptions.  This conflict set tells us that, as long as
all these conflict assumptions are kept, the instance will remain
UNSAT... i.e., these conflict assumptions suffice to prevent the
partial assignment from overlapping with any nogood.

We will also find our Tseitin output variable in the conflict set, but
that's OK: we don't want to change that variable's value, so we can
ignore it.