.## Check and reduce assignments with Tseitin
Given a partial assignment `a`, we want to know if there exists
an extension `x` of `a` such that:

1. `x` satisfies the domain constraints
2. `x` violates at least one of the nogoods

Really, we want to know when that's impossible, i.e., forall
valid extensions `x` of `a`, `x` does not violate any nogood,
but SAT works with proofs of existence.

The Tseitin solver contains a set of clauses which guarantees that the
current "output" variable will be true iff `x` violates one of the
nogoods.

We can thus check if a given partial assignment is disjoint from the
nogoods by asking the Tseitin SAT solver to find us an
extension (solve with the assumptions that `output = true`, plus all
the partial assignment `a` itself).  If it finds one, we know the
assignment is not disjoint.

If `CryptoMiniSat` proves that there is no such extension (i.e., the
partial assignment is disjoint), it will also return a learned clause,
derived from the clauses in the solver (but not the assumptions) such
that the learned clause contradicts the assumptions.

Every literal in the learned (CNF) clause appears in the
assumptions (otherwise there would be some degree of freedom to avoid
a contradiction), and is the negation of the assumption: if `x = true`
is in the learned disjunctive clause and in the assumptions, the
learned clause is satisfied by `x = true`, and there is no
contradiction.

Anything in the assumptions that does not appear in the learned clause
is redundant for disjointedness: our partial assignment is disjoint
from the nogoods as long as the Tseitin system is UNSAT, and any
assignment that contradicts the learned clause implies UNSAT.  We can
thus drop assumption that does not appear in the learned clause.

Now, the learned clauses are a side-effect of the CDCL solving
process, and there is no guarantee on their quality.  Of course, we
can't expect a minimum conflict: that's too hard.  Perhaps
surprisingly, we're not guaranteed to have a minimal conflict either!

If we really want to reduce the partial assignment while preserving
disjointedness, we want to probe every variable in the assignment.
We'll remove that variable from the assumptions, and see if the result
is still UNSAT: if so, we don't need the variable, and we can use the
learned clause to detect that some other variables are also redundant.
If the system is not satisfiable, we know removing the variable loses
disjointedness, so we add it to a set of mandatory variables.

Once we've probed every variable in the assumptions (or dropped them
due to learned clauses), the set of mandatory variables is a minimal
disjoint partial assignment: removing anything from that list
introduces some overlap between the nogoods and the assignment's
extensions.

One might have a really good idea about the "best" order in which to
probe variables, maybe as part of some GRASP heuristic.  That's why we
also offer a fully in-order reduction mode, where we only use learned
conflict clauses when they don't jump the probing queue.