.# The propagation engine
The core of a coherent logic solver is the propagation engine, which
implements the few (higher-order) deduction rules that define coherent
logic.

Coherent logic is a subset of first-order logic, so it only
manipulates variables, opaque objects variables with identity but not
otherwise imbued with any meaning, and predicates on tuples of
variables.  These predicates represent things we know or can assume to
be true, and absence *is not* negation (i.e., we have to assume an
[open world](https://en.wikipedia.org/wiki/Open-world_assumption)).

The [datalog](https://en.wikipedia.org/wiki/Datalog) programming
language successfully implements the correspondence between logical
predicate and "tables" in databases. We'll do the same in TCL, and, in
fact, most of the code in the propagation engine actually implements
datalog!

We implement TCL's propagation engine (`timely-coherent-propagator`)
on top of [differential dataflow](https://docs.rs/differential-dataflow)
because we expect to frequently make changes to the input
state (assumptions); in particular, I'm betting that free-er form
traversals than backtracking will make it easier to implement hybrid
semi-automated search strategies.