.## Unification
Now that we know how to represent `Variable`s and `Fact`s (tuples of
variables), the next step is matching `Fact`s against patterns to
generate `Capture`s, and, in the opposite, instantiate patterns with
`Capture`s to generate `Fact`s.

In order to do all that, we need a data structure describe patterns.
We're only concerned with tuples of `Variable`s, so that's an easy
task: there's no nesting, or variation in length.

We identified that what we're really interested in is matching `Fact`s
to generate `Capture`s, and converting `Capture`s to `Fact`s.  At the
heart of all that is manipulating environments, mappings from
pattern `MetaVar`iable to `Variable`s.

We will store these environments in DD collections, so we don't want
nor need to represent them as associative data structures like hash
tables or alists: we can maintain a `MetaVar`-to-index mapping next to
each collection, and store each environment as a `Capture` record.

This means that we will also have to explicitly shuffle `Capture`s
around to change their shape (e.g., to make sure two collections of
captures are compatible before merging them).

That's why we'll start with defining pattern `MetaVar`iables and
generating shuffle functions at runtime.