.## The `execution` module
Differential dataflow is statically and (heavily generically) typed.
That's not a great fit for us, since we don't know our query plan at
compile-time... and we mostly work with collections of `Fact`s or
`Capture`s.

The execution module wraps DD's functionality in a runtime type
system, where the collections and operators are checked for
compatibility eagerly, as the dataflow graph is constructed (and not
when operating on individual datum).

This includes shape-annotated wrappers around DD containers and
`sinks` to access a collection's values, as well as code to lower join
plans into differential dataflow graphs.