.## A trivial and naïve Datalog engine
Now that we know how to match antecedents in coherence logic
sequences, we can implement a datalog engine to apply all rules that
fit the datalog model until saturation.  Such "trivial" rules are all
the rules without disjunctions nor existential in the consequent: in
order to apply these rules, we don't have to pick what pattern to
instantiate, nor generate a new variable.

We already have the matching half of a bottom-up datalog engine.  We
simply have to convert `Capture`s to `Fact`s, and to tie the knot for
DD to compute a fixpoint for us.