.## Matching full-blown sequents
The datalog engine handles trivial sequents.  We now have to do
something about sequents that make us choose what to derive from a
given left-hand side `Capture`, or sequents with existentials in the
consequents (i.e., consequents that introduce new variables).

We'll defer these decisions to higher-level modules than the
propagator; we simply want to report the list of all decisions there
are to be made, i.e., generate a collection of matched `Capture`s for
each sequent.

And this "matching" of non-trivial rules is where we implement
something subtle and specific to coherent logic: whenever a sequent's
consequents are already satisfied for a given capture, we must do
nothing.  The benefit is clear: we can cut down on a lot of useless
instantiations.  What's more surprising is that we can still prove
a lot of theorems with this restriction.

What do we specifically mean by "consequents are already satisfied?"
Consequents are disjunctions of conjunctions; if even one conjunction
is already known to be true, we don't want to instantiate anything.
Logically, this makes sense: if we have $x \vee y$ as the right-hand
side of an implication, but we already know $x$, the implication is
vacuous.

We know a conjunction is true when the result of instantiating it with
the candidate `Capture` is already in our collections of facts.  It's
a bit subtle when there's an existential: a consequent like
$\exists y. p(x, y)$ is satisfied whenever we can find a match for
$p(x, \cdot)$, i.e., whenever $\forall y. p(x, y)$ yields a non-empty set
of matches.

This is starting to sound very similar to our left-hand side matching
logic.  In fact, if we pretend the $\exists$ are $\forall$, it's the
exact same thing!  The only difference is that we may have multiple
such right-hand side joins for a single sequent, since a sequents's
consequents form a *disjunction*.

For each such right-hand side joins, we will look at the matched
`Capture`s and *anti*-join them from the left-hand side `Capture`s,
to disable any `Capture` that would yield a right-hand side that
is already known.