backend: probe each member of a partial assignment when reducing
When the SAT solver gives us a conflict clause, we definitely know
that literals that do not appear in the conflict can be removed from
the partial assignment.  However, there is no guarantee that literals
that do appear in the conflicts are truly necessary.  Probe each
literal before keeping it: if we find overlap when we remove the
literal, then we really want to keep it.